2020_85rec
Given a matrix filled with values of 1, 0, and -1, the task is to return the maximum points that can be accumulated along the longest path in the matrix. The movement rules are as follows:

Odd-numbered rows: You can move left and down.
Even-numbered rows: You can move right and down.
The goal is to find the path that yields the highest sum of values while adhering to these movement restrictions.